Implement abstract class for balance entities
